Change AreYouSureView to be a util
