+++ BOT/lib/cogs/product.py
@@ -14,6 +14,7 @@ from nextcord.ui.select import select
 from nextcord.user import BU
 from ..utils.api import *  # Imports everything from the API util
 from ..utils.database import find
+from ..utils.util import AreYouSureView
 import json
 
 productoptions = []
@@ -34,62 +35,62 @@ class CancelView(ui.View):
 
 
 # Are you sure?
-class AreYouSureView(ui.View):
-    def __init__(self, context):
-        super().__init__(timeout=None)
-       # self.Action = Action
-       # self.args = args
-        self.context = context
-        self.Return = None
+# class AreYouSureView(ui.View):
+#     def __init__(self, context):
+#         super().__init__(timeout=None)
+#        # self.Action = Action
+#        # self.args = args
+#         self.context = context
+#         self.Return = None
 
-    @ui.button(
-        label="Yes", custom_id="products:yes_I_am_sure", style=ButtonStyle.success
-    )
-    async def iamsure(self, _, interaction: Interaction):
-        self.Return = True
-        # if self.Action == "deleteproduct":
-        #     try:
-        #         deleteproduct(self.args[0])
-        #         await interaction.message.delete()
-        #         await interaction.response.send_message(
-        #             f"Deleted {self.args[0]}.",
-        #             ephemeral=True,
-        #         )
-        #         self.stop()
-        #     except:
-        #         await interaction.message.delete()
-        #         await interaction.response.send_message(
-        #             f"Failed to delete {self.args[0]}.",
-        #             ephemeral=True,
-        #         )
-        #         self.stop()
-        # if self.Action == "updateproduct":
-        #     try:
-        #         updateproduct(
-        #             self.args[0], self.args[1], self.args[2], self.args[3], self.args[4]
-        #         )
-        #         await interaction.message.delete()
-        #         await interaction.response.send_message(
-        #             f"Updated {self.args[0]}.",
-        #             ephemeral=True,
-        #         )
-        #         self.stop()
-        #     except:
-        #         await interaction.message.delete()
-        #         await interaction.response.send_message(
-        #             f"Failed to update {self.args[0]}.",
-        #             ephemeral=True,
-        #         )
-        self.stop()
+#     @ui.button(
+#         label="Yes", custom_id="products:yes_I_am_sure", style=ButtonStyle.success
+#     )
+#     async def iamsure(self, _, interaction: Interaction):
+#         self.Return = True
+#         # if self.Action == "deleteproduct":
+#         #     try:
+#         #         deleteproduct(self.args[0])
+#         #         await interaction.message.delete()
+#         #         await interaction.response.send_message(
+#         #             f"Deleted {self.args[0]}.",
+#         #             ephemeral=True,
+#         #         )
+#         #         self.stop()
+#         #     except:
+#         #         await interaction.message.delete()
+#         #         await interaction.response.send_message(
+#         #             f"Failed to delete {self.args[0]}.",
+#         #             ephemeral=True,
+#         #         )
+#         #         self.stop()
+#         # if self.Action == "updateproduct":
+#         #     try:
+#         #         updateproduct(
+#         #             self.args[0], self.args[1], self.args[2], self.args[3], self.args[4]
+#         #         )
+#         #         await interaction.message.delete()
+#         #         await interaction.response.send_message(
+#         #             f"Updated {self.args[0]}.",
+#         #             ephemeral=True,
+#         #         )
+#         #         self.stop()
+#         #     except:
+#         #         await interaction.message.delete()
+#         #         await interaction.response.send_message(
+#         #             f"Failed to update {self.args[0]}.",
+#         #             ephemeral=True,
+#         #         )
+#         self.stop()
 
-    @ui.button(
-        label="No", custom_id="products:no_I_am_not_sure", style=ButtonStyle.danger
-    )
-    async def noiamnotsure(self, _, interaction: Interaction):
-        self.Return = False
-        # await interaction.message.delete()
-        # await interaction.response.send_message("Canceled action.", ephemeral=True)
-        self.stop()
+#     @ui.button(
+#         label="No", custom_id="products:no_I_am_not_sure", style=ButtonStyle.danger
+#     )
+#     async def noiamnotsure(self, _, interaction: Interaction):
+#         self.Return = False
+#         # await interaction.message.delete()
+#         # await interaction.response.send_message("Canceled action.", ephemeral=True)
+#         self.stop()
 
 
 # Delete view
