+++ BOT/lib/utils/util.py
@@ -4,12 +4,34 @@
 """
 
 from quart import request
+from nextcord import ui, Interaction, SelectOption, ButtonStyle
 import json
 import functools
 
 with open("./BOT/lib/bot/config.json") as config_file:
     config = json.load(config_file)
 
+
+# Are you sure?
+class AreYouSureView(ui.View):
+    def __init__(self, context):
+        super().__init__(timeout=None)
+        self.context = context
+        self.Return = None
+
+    @ui.button(
+        label="Yes", custom_id="products:yes_I_am_sure", style=ButtonStyle.success
+    )
+    async def iamsure(self, _, interaction: Interaction):
+        self.Return = True
+        self.stop()
+
+    @ui.button(
+        label="No", custom_id="products:no_I_am_not_sure", style=ButtonStyle.danger
+    )
+    async def noiamnotsure(self, _, interaction: Interaction):
+        self.Return = False
+        self.stop()
 
 def require_apikey(view):
     # Makes it so I dont repeat if apikey in every website base.
